+++ lib/rubocop/cop/performance/casecmp.rb
@@ -18,7 +18,7 @@ module RuboCop
         MSG = 'Use `casecmp` instead of `%s %s`.'.freeze
 
         def_node_matcher :downcase_eq, <<-END
-          (send $(send _ ${:downcase :upcase}) ${:== :eql?} _)
+          (send $(send _ ${:downcase :upcase}) ${:== :eql? :!=} _)
         END
 
         def on_send(node)
@@ -29,7 +29,7 @@ module RuboCop
         end
 
         def autocorrect(node)
-          receiver, _method, arg = *node
+          receiver, method, arg = *node
           range = Parser::Source::Range.new(node.source_range.source_buffer,
                                             receiver.loc.selector.begin_pos,
                                             arg.loc.begin.begin_pos)
@@ -48,6 +48,7 @@ module RuboCop
               corrector.replace(range, 'casecmp(')
               corrector.insert_after(arg.source_range, ').zero?')
             end
+            corrector.insert_before(receiver.source_range, '!') if method == :!=
           end
         end
       end