+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Changes
 
 * Remove `Style/SpaceAfterControlKeyword` and `Style/SpaceBeforeModifierKeyword` as the more generic `Style/SpaceAroundKeyword` handles the same cases. ([@lumeet][])
+* Handle comparisons with `!=` in `Performance/casecmp`. ([@segiddins][])
 
 ## 0.36.0 (14/01/2016)